requirements.txt
### requirements.txt
```
Pillow
numpy
```